requirements.txt
flask==3.1.2
eventlet==0.40.3
python-socketio==5.14.2
paramiko==4.0.0